+++ pyanalyze/arg_spec.py
@@ -65,6 +65,7 @@ from typing import (
     Any,
     Callable,
     Iterator,
+    List,
     Sequence,
     Generic,
     Mapping,
@@ -589,7 +590,13 @@ class ArgSpecCache:
                 )
 
         allow_call = FunctionsSafeToCall.contains(obj, self.options)
-        argspec = self.ts_finder.get_argspec(obj, allow_call=allow_call)
+        if safe_isinstance(obj, (type, str)):
+            type_params = self.get_type_parameters(obj)
+        else:
+            type_params = []
+        argspec = self.ts_finder.get_argspec(
+            obj, allow_call=allow_call, type_params=type_params
+        )
         if argspec is not None:
             return argspec
 
@@ -749,8 +756,17 @@ class ArgSpecCache:
             # Python 2.
             return None
 
+    def get_type_parameters(self, typ: Union[type, str]) -> List[Value]:
+        bases = self.get_generic_bases(typ, substitute_typevars=False)
+        tv_map = bases.get(typ, {})
+        return [tv for tv in tv_map.values()]
+
     def get_generic_bases(
-        self, typ: Union[type, str], generic_args: Sequence[Value] = ()
+        self,
+        typ: Union[type, str],
+        generic_args: Sequence[Value] = (),
+        *,
+        substitute_typevars: bool = True,
     ) -> GenericBases:
         if (
             typ is Generic
@@ -766,14 +782,15 @@ class ArgSpecCache:
         if not my_typevars:
             return generic_bases
         tv_map = {}
-        for i, tv_value in enumerate(my_typevars.values()):
-            if not isinstance(tv_value, TypeVarValue):
-                continue
-            try:
-                value = generic_args[i]
-            except IndexError:
-                value = AnyValue(AnySource.generic_argument)
-            tv_map[tv_value.typevar] = value
+        if substitute_typevars:
+            for i, tv_value in enumerate(my_typevars.values()):
+                if not isinstance(tv_value, TypeVarValue):
+                    continue
+                try:
+                    value = generic_args[i]
+                except IndexError:
+                    value = AnyValue(AnySource.generic_argument)
+                tv_map[tv_value.typevar] = value
         return {
             base: {tv: value.substitute_typevars(tv_map) for tv, value in args.items()}
             for base, args in generic_bases.items()